requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
pydantic
PyPDF2
python-docx
fpdf2
